Added UCSC API functions, utility functions and constants to new .py files
