+++ requirements.txt
@@ -1,0 +1,7 @@
+numpy>=1.23,<2
+pandas>=1.5,<2
+matplotlib>=3.5,<4
+seaborn>=0.12,<1
+requests>=2.28,<3
+mysql-connector-python>=8.0,<9
+Pillow>=9.0,<10
